// apps/main/src/models/v2/iam/userWithRole.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { SubjectType } from './subjectType';
import type { UserWithRoleType } from './userWithRoleType';

/**
 * A user / service user with their role
 */
export interface UserWithRole {
  /** The user / service user id */
  id: string;
  /** The type of user - human user or service user */
  type: UserWithRoleType;
  /** The membership subject type. */
  subject_type: SubjectType;
  /** The membership subject id. */
  subject_id: string;
}